Allow servers control over deletion
Before, when a client called a destructor the generated handlers
automatically removed it and replied with a confirmation. However,
for proxy servers it's convenient to delay the confirmation until
the upstream proxy is also deleted. This way, we can continue proxing
events from the upstream object and the client doesn't see anything
unusual.